Fix union and collection validation edge cases
Some things passed/returned to Thrift endpoints are not themselves
Thrift objects, e.g., lists, primitives, or nil. These can and should
still be validated. Since we're anyway not making use of type hints in
the schema and relying on Thrift structs to self-validate, we can just
recurse the struct tree and let what can be validated be validated,
greatly cleaning up implementation while adding functionality.

For example, union validation was broken for a very specific edge case:
when a field that was a container around structs is not set, validation
produces a highly confusing error message indicating that the unset
field is, in fact, not set. Consider the Thrift definition

  union UnionExample {
    1: list<SomeStruct> primary,
    2: string alternate,
  }

Setting the `alternate` field caused validation to fail, claiming that
the `primary` field was not the set_field. This was due to the validator
accessing the unset field, because the code meant to check this case
only ran for specific field types. Now it performs correctly regardless
of field type.