Use rake for make tasks
We already need `rake` to work with this gem, and it's capable of
basically everything that `make` is; unify build tools to prevent
confusion.

Move the test thrift files into the test/ directory to avoid clutter,
plus add a `clean` target for `rake` that wipes built definitions.

Also, add minitest as a dependency, since it’s not built into newer
Ruby versions.